無線(gps計測)
