+++ nakamura/無線/main_gps.py
@@ -1,13 +1,19 @@
 import time
 import gps
+import radio
 
 class cansat(object):
     def __init__(self):
         self.gps = gps.GPS()
+        self.radio = radio.radio("/dev/ttySOFT0", 15)
 
     def setup(self):
-        self.gps.setupGPS()
-        
+        #self.gps.setupGPS()
+        self.radio.setup()
+
+    def main(self):
+        self.radio.sendData()
+    """
     def writeData(self):
         self.gps.gpsread()
         timer = 1000*(time.time() - start_time)
@@ -19,10 +25,15 @@ class cansat(object):
     
         with open("test.txt",mode = 'a') as test:
             test.write(datalog + '\n')
+            """
 
-
-start_time = time.time()
 cansat = cansat() 
 cansat.setup()
+#cansat.main()
+
 while True:
-    cansat.writeData()   
+    try:
+        cansat.main()
+    except KeyboardInterrupt:
+            self.sendDevice.close()
+
