+++ nakamura/無線/modules.order
@@ -1,0 +1,1 @@
+kernel//home/pi/Desktop/0620/soft_uart.ko
