+++ nakamura/無線/soft_uart.mod.c
@@ -73,4 +73,4 @@ __attribute__((section(".modinfo"))) =
 "depends=";
 
 
-MODULE_INFO(srcversion, "FF97B9BF705FF2E8754580E");
+MODULE_INFO(srcversion, "4B13341D35424E3C10758C5");
