--- nakamura/無線/soft_uart/modules.order
@@ -1,1 +1,0 @@
-kernel//home/pi/soft_uart/soft_uart.ko
